Удаляет компонент PageScreen
