--- src/components/page-screen/page-screen.test.js
@@ -1,47 +1,0 @@
-import React from "react";
-import renderer from "react-test-renderer";
-import PageScreen from "./page-screen";
-import {EMPTY_STRING, Value} from "../../constants";
-
-describe(`Make snapshot`, () => {
-  const initProps = {
-    films: [{
-      id: Value.EMPTY,
-      name: `Keks`,
-      image: {
-        poster: EMPTY_STRING,
-        posterAlt: EMPTY_STRING,
-        preview: EMPTY_STRING,
-        previewAlt: EMPTY_STRING,
-        background: EMPTY_STRING,
-        backgroundAlt: EMPTY_STRING,
-      },
-      backgroundColor: EMPTY_STRING,
-      video: {
-        link: EMPTY_STRING,
-        preview: EMPTY_STRING,
-      },
-      description: EMPTY_STRING,
-      rating: Value.FULL,
-      scoresCount: Value.FULL,
-      director: `Keks`,
-      starring: [`Me`, `You`, `They`],
-      runTime: Value.EMPTY,
-      genre: EMPTY_STRING,
-      released: Value.FULL,
-      isFavorite: true,
-    }],
-    genres: [`ECMA`, `CSS`],
-    icons: [`PAUSE`],
-  };
-
-  it(`Get snapshot component`, () => {
-    const tree = renderer
-      .create(
-          <PageScreen {...initProps}/>
-      )
-      .toJSON();
-
-    expect(tree).toMatchSnapshot();
-  });
-});
